+++ src/components/molecules/Visualizer/Engine/Cesium/PhotoOverlay/hooks.ts
@@ -5,7 +5,7 @@ import { useCallback, useEffect, useRef } from "react";
 import { useDelayedCount, Durations } from "@reearth/util/use-delayed-count";
 import { Camera } from "@reearth/util/value";
 
-import { useVisualizerContext } from "../../../context";
+import { useVisualizerContext } from "../../../Plugin";
 
 export type { TransitionStatus } from "@rot1024/use-transition";
