+++ src/components/molecules/Visualizer/Widget/index.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, Story } from "@storybook/react";
 import { Math as CesiumMath } from "cesium";
 import React from "react";
 
-import { Provider } from "../context";
+import { Provider } from "../Plugin";
 import { context } from "../storybook";
 
 import Widget, { Props } from ".";
